requirements.txt
django
django-bootstrap-v5
django-widget-tweaks
psycopg2
psycopg2-binary
djangorestframework